unify receive form for all networks
