--- packages/suite/src/views/wallet/account/receive/bitcoin/messages.ts
@@ -1,10 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-const definedMessages = defineMessages({
-    TR_RECEIVE_BITCOIN: {
-        id: 'TR_RECEIVE_BITCOIN',
-        defaultMessage: 'Receive Bitcoin',
-    },
-});
-
-export default definedMessages;
