--- packages/suite/src/views/wallet/account/receive/ethereum/messages.ts
@@ -1,10 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-const definedMessages = defineMessages({
-    TR_RECEIVE_ETHEREUM_OR_TOKENS: {
-        id: 'TR_RECEIVE_ETHEREUM_OR_TOKENS',
-        defaultMessage: 'Receive Ethereum or tokens',
-    },
-});
-
-export default definedMessages;
