--- packages/suite/src/views/wallet/account/receive/ripple/messages.ts
@@ -1,10 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-const definedMessages = defineMessages({
-    TR_RECEIVE_RIPPLE: {
-        id: 'TR_RECEIVE_RIPPLE',
-        defaultMessage: 'Receive Ripple',
-    },
-});
-
-export default definedMessages;
